+++ src/app/common.blocks/modal/PreferencesModal.scss
@@ -8,13 +8,13 @@
   hr {
     margin: 16px 0;
   }
-}
 
-.PreferencesModal-Title {
-  margin-top: 16px;
-  margin-bottom: 16px;
-}
+  .PreferencesModal-Title {
+    margin-top: 16px;
+    margin-bottom: 16px;
+  }
 
-.PreferencesModal-Title:first-child {
-  margin-top: 0;
+  .PreferencesModal-Title:first-child {
+    margin-top: 0;
+  }
 }
