+++ src/app/component.blocks/CanvasTooltip.tsx
@@ -2,9 +2,10 @@ import { action } from "mobx";
 import { Tooltip, TooltipProps, Typography, styled, tooltipClasses } from "@mui/material";
 import { getAppStores } from "@core/MainApp";
 import classNames from "classnames";
+import "./CanvasTooltip.scss";
 
 export const Padding0Tooltip = styled(({ className, ...props }: TooltipProps) => (
-  <Tooltip {...props} classes={{ popper: classNames(className, "canvas-tooltip") }} />
+  <Tooltip {...props} classes={{ popper: classNames(className, "CanvasTooltip") }} />
 ))(({ theme }) => ({
   [`& .${tooltipClasses.tooltip}`]: {
     padding: "0",
@@ -12,14 +13,14 @@ export const Padding0Tooltip = styled(({ className, ...props }: TooltipProps) =>
   }
 }));
 
-export const Label = function (props: { text: string; onClick: () => void }) {
+export const CanvasTooltip = function (props: { text: string; onClick: () => void }) {
   const { app } = getAppStores();
 
   return (
     <Typography
       variant="body2"
       component="span"
-      className="canvas-tooltip-label"
+      className="CanvasTooltip-Label"
       onClick={action(() => {
         props.onClick();
         app.fieldEditor.tooltipPosition = undefined; // UX: Hide tooltip
